bot/requirements.txt
requests==2.28.1
pandas==1.5.0
antlr4-python3-runtime==4.9.3
tabulate